pyproject.toml
[project]
name = "pico-202510"
version = "0.1.0"
description = "pico-python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.1.0",
    "django>=5.0.0",
    "paho-mqtt>=2.0.0",
    "streamlit>=1.51.0",
]
